Skip downloading existing azlux.fr GPG key

+++ .config/setup/installers/apt.zsh
@@ -7,11 +7,15 @@ move_legacy_apt_keys() {
     log "Updating APT keys..." debug
     sudo mkdir -p "$new_keyring_dir"
 
-    # Download and add the new azlux.fr GPG key
-    if curl -fsSL https://azlux.fr/repo.gpg.key | sudo gpg --dearmor -o "$new_keyring_dir/azlux.gpg" 2>/dev/null; then
-        log "Downloaded and added new azlux.fr GPG key." debug
+    # Download and add the new azlux.fr GPG key if it doesn't exist
+    if [[ ! -f "$new_keyring_dir/azlux.gpg" ]]; then
+        if curl -fsSL https://azlux.fr/repo.gpg.key | sudo gpg --dearmor -o "$new_keyring_dir/azlux.gpg" 2>/dev/null; then
+            log "Downloaded and added new azlux.fr GPG key." debug
+        else
+            log "Failed to download or add new azlux.fr GPG key." error
+        fi
     else
-        log "Failed to download or add new azlux.fr GPG key." error
+        log "azlux.fr GPG key already exists. Skipping download." debug
     fi
 
     # Update the azlux.fr repository configuration
